This is the first version of the project to actually be fully playable all the way through, granted using buttons.  You can play the game, start to finish, however, it does not terminate at the end of a game, nor does it allow you to start a new game.   The game itself however is now in a playable state, just in need of tweaking.  Also, removed some unused code I had made to display the command line version I first committed, as it is useless in this version.  There are other cleanups still to do with unused code. The game will need to be tested on other emulators still.  It has been successfully run only on a Nexus One running API 18, but that is all the testing PC I have can operate effectively.
Todo
End game situation, along with a score and deal that updates as the game is played.  An updated title screen and instructions on how the game is played.  Hopefully card animations and sounds as well.

There is a listed error, the activity_game.xml file shows an error because it believes layout_width and layout_heidth should be defined, however, the percent support library allows using layout_widthPercent and layout_heightPercent in place of these.  The error listed for those can be ignored.