Updated the title screen and the end game so that there would be the ability to play again, though it is necessary to go back to the title screen. (Allowing a new game instead should be simple, but is not critical at this time.  Also added in the ability to go to a new screen to view a how to play instruction from the title menu, which goes back to the title menu upon button click.  Added a new class for this instruction page, as it is a new activity, and required adding a new xml layout, activity_instructions.xml as well to be tracked.  Removed unused code relics from the linux version of the program, and removed the xml files that were not related to the project outside being used for experimentation purposes (activity_idsolitaire2.xml, title_screen.xml).  Added in copyrights notices to the top everywhere appropriate, or moved if already there but not at the top.
Done: End game situation,  An updated title screen and instructions on how the game is played.

Todo: Hopefully card animations and sounds as well. A score and cards remaining that updates as the game is played

There is a listed error, the activity_game.xml file shows an error because it believes layout_width and layout_heidth should be defined, however, the percent support library allows using layout_widthPercent and layout_heightPercent in place of these.  The error listed for those can be ignored.